Change syntax for generic parameters.
Using '<' and '>' to denote generic type parameters (as is common
in other languages) quickly became impractical because it is
ambiguous with the "less than" operator. It is technically
possible to implement the syntax this way, but I could not figure
out a way to do it with the one token look-ahead scanner that
ships with Go (I believe you would need arbitrary look-ahead).

The new syntax involves a new token (token.DOUBLE_COLON) which
removes any ambiguity and is far easier to parse. I just had to
make a small change to the way errors are reported in slice
indexes, and other than that nothing appears to have broken.

So far, the parser supports struct type definitions and struct
literals with generic type parameters.