Create new ConcreteTypeParamList type and use it where appropriate.
This allows us to differentiate between generic type parameters in
a declaration (in which case we expect the list to contain plain
identifiers) and concrete type parameters being provided to such
declarations (in which case we expect concrete types). As a result,
we get improved errors in the parser. We also will be able to parse
and transform more complicated concrete types in the future, e.g.,
selector expressions.