Move type parameters to identifier in type and func declarations.
This move greatly simplifies the parser and type-checker while also
allowing for more flexibility in the kinds of types which can be
generic. Instead of only structs, any named type can now be generic
and have type parameters.

Although this may be subjective, the new syntax also comes across
as more readable since the type parameters in usage and declaration
are both now written in the same place: after the identifier name.

Finally this change also makes some restrictions of the current
iteration of the language more clear. Namely, we currently don't
allow generic anonymous types. With the new syntax, it is a syntax
error to even try to write such types (whereas before, one might
try adding type parameters e.g., after the 'struct' key word for
an anonymous struct and the parser would accept it).