Implement a new bracket syntax for generic type parameters.
This new syntax will be more familiar to developers (e.g. Scala
uses a similar syntax), slightly easier to type, and less strange/
symbolic for readers. In addition, it allows us to remove the
DCOLON token entirely.

The main downside of the new syntax is that it introduces ambiguity
in the parser. In most cases, this ambiguity can be resolved by
looking ahead a few tokens. In some cases, the ambiguity cannot be
resolved by the parser and must be resolved later on by the type-
checker. This is not so different from how the existing Go
toolchain handles ambiguity between function calls and type
conversions. Critically, the ambiguity has little impact on the
overall structure of the AST and is not too difficult to address
in the type-checker and transformer.