Rename some things for greater clarity.
Specifically, in order to distinguish between declaration and
usage of generic types, we now use the terms "type parameters" and
"type arguments". Type parameters are declared names for arbitrary
types and appear in the declaration of functions and named types.
Type arguments are concrete types supplied as arguments when a
generic function or named type is used.