Allow usage of certain generic receiver types without type params.
For example, if a generic type `type A[T] []T` is used as a
receiver without type params, the type checker will still allow
indexing and ranging over it. (We know that regardless of the
concrete type, certain things are gauranteed to work). It also
works if the underlying type is an array, channel, function, etc.