Fix bugs having to do with generic methods.
One bug occurred when a method with a generic receiver also had
generic type parameters. The method was not being transformed
correctly in some cases.

The other had to do with methods with the same name but different
receiver types. They were not being transformed correctly because
one declaration would overwrite the other.