Report error for generic functions as values without type args.
Technically, the type-checker already supports generic
functions/methods as values, but the transformer does not. The
transformer is currently only capable of transforming expressions
involving the original function name and when functions are used
as values, the name could be completely different. We don't
currently have a way to track usage and perform the necessary AST
modifications when the names are different.